Added species from 2 major pathways directly into the core
According to 10.1016/j.combustflame.2013.04.023, they should be direct products
from iso-butanol, but were not formed. The isobutanol radical that was not formed was
particularly concerning, and needs to be further investigated. It should be the major product
from iso-butanol, but wasn't included in the core well into the run. What strategies would
get this in the core earlier?

The other 3 major pathways we in the model core.